Make gunicorn use STATSD_* env vars
Make gunicorn use the same STATSD_* environment variables as bouncer
itself uses, for the metrics that it reports to statsd.

Add a gunicorn.conf.py file for retrieving these environment variables
and falling back to the same defaults that bouncer does.

Have the Dockerfile (for production) and the Makefile (for dev) use the
new gunicorn.conf.py file.

Remove the STATSD_MAXUDPSIZE environment variable from the README.
Gunicorn doesn't support this setting, so it would only work for the
metrics reported by bouncer.