Don't use append_slash=True
Tweak the routes so that the following three will all work:

    /<annotation_id>
    /<annotation_id>/
    /<annotation_id>/<path>

The first one just works, it does _not_ redirect to the one with the
trailing slash.

The reason for this is that with append_slash=True an HTTPNotFound is
raised, caught by the Raven tween and reported to Sentry, and _then_
ends up in the notfound view which redirects to the same URL but with a
trailing slash.

So a 404 would uselessly be reported to Sentry each time a URL without a
trailing slash was visited.