Add Sentry reporting of JavaScript crashes
Add raven.min.js (Sentry's JavaScript client) to the annotation
redirector page (currently just using Sentry's CDN-hosted copy).

Add an additional environment variable SENTRY_JAVASCRIPT_DSN since the
Sentry DSN that the client-side JavaScript code uses is different to the
one that the server-side code uses (it doesn't contain the secret key
part because it isn't secure).

The server renders the SENTRY_JAVASCRIPT_DSN into the HTML along with
the rest of the settings.

The JavaScript code in redirect.js picks this up and uses it to
configure Raven, which automatically reports crashes to Sentry.

Also add the bouncer release version number to the settings object that
the server sends to the JavaScript so that JavaScript crash reports can
include the version number as Python ones do.