Serve static URLs with same protocol as the site
The CSS in the bouncer production deployment isn't being loaded because
of a mixed-content warning. Pyramid's static_url is generating http://
URLs for static assets, and bouncer is deployed with https://.

Use static_path instead of static_url everywhere. This generates paths
(relative URLs) instead of full URLs, so they should use the same
protocol as the request.