Add environment name to Sentry client
Skyliner provides the ENV environment variable for our application,
which contains either "prod" or "qa" depending on the environment.

Sentry supports partitioning reports based on the environment within a
single application. Adding this metadata to the `Client` allows us to
migrate QA and Prod onto a single application in Sentry.