Add `/go` endpoint for visiting a URL with annotations.
Add a `/go?url={url}` endpoint which takes the user to a given URL with
the annotation layer enabled, optionally with the client configured to
initially show only annotations matching a given filter by specifying a
"q={query}" parameter.

The {query} is passed directly to the search input in the client.

For the API design rationale, see [1] and [2].

[1] https://gist.github.com/robertknight/85c9dfc89910288e84588903ab1010a3

[2] https://groups.google.com/a/list.hypothes.is/forum/#!topic/dev/sFH-3xmeK8Y