Do not redirect user to proxy for annotations created by third-party users
Annotations in groups belonging to third party authorities can only be
shown using the embedded client on the original URL, not the Via proxy
because configuration from the host page is needed to make the client
show the correct group and log the user in to the third-party managed
account.

Since we know for these annotations that the target URL must embed the
client, we can always redirect the user directly to it, rather than
using the proxy.