Avoid redirecting user to Via if target URL is known to embed client
If the target URL is known to embed the client, sending the user
directly to that URL will be a better experience than sending the user
through a proxy service.

There are various ways that we could detect whether a target URL embeds
the client. This commit implements the dumbest possible approach of a
hardcoded list of URL patterns.

For consistency with how annotation search works, HTTP and HTTPS are
treated as equivalent. The query string and fragment are also ignored
when testing for a possible match.