Refactor Makefile and tox.ini
Refactor Makefile and tox.ini to be the same as the ones used in h and
soon to be used in all our apps.

Also add a .coveragerc file, the same as h has, instead of passing args
to it on the command line.

`requirements-dev.txt` is removed as its no longer needed. There weren't
actually any "dev" (as in running the dev server) requirements in here,
they were all requirements for running the tests and lint, and those are
listed in `tox.ini` instead now.