Add the ability to focus a group via querystring parameter
For the Community Groups project, we need to be able to
generate bouncer links for a URL that automatically
focus a specific group (by group pubid) in the client
layer.

The client has been updated to understand an
extended fragment syntax (`#annotations:group:{pubid}`).
This adds the logic such that bouncer can accept a
`group` query parameter on the `/go` route and
generate a redirect with the correct fragment
syntax for focusing a group.