Use "activate" API to redirect and activate extension
If the installed Hypothesis Chrome extension supports the "activate" message,
use it to redirect the current tab and activate the extension.  This enables
bouncer to activate just one specific Hypothesis extension, if there are
multiple builds installed. The previous method would result in all installed
builds trying to activate.

In the process of using this API, the `redirect` function was converted
to be async, so that it can report when all of its async internal steps
are done. This makes writing the tests much easier. The tests have also
been rewritten to more clearly reflect the meaningfully different
scenarios that can occur.

Part of https://github.com/hypothesis/browser-extension/issues/1177