Support setting a different extension ID per authority
Allow the Chrome extension ID to be customized depending on the authority that
an annotation is associated with. The extension IDs are specified by setting the
`CHROME_EXTENSION_ID` env var to a JSON map of authority to extension ID. The
map must contain a "default" field which specifies the extension to use if no
other entry matches.

For example:

```
export CHROME_EXTENSION_ID='{"default":"abc...", "atlassian.hypothes.is": "def..."}'
```