Redirect user to Via if annotation is a transcript annotation
When following a bouncer link for an annotation, check if the annotation is a
YouTube transcript annotation made using Via. If so, redirect the user to Via
even if they have the extension installed.

Via's video player app does not currently record a specific marker indicating
that it was used to create the annotation, but it is the only tool that will
create an annotation with a `MediaTimeSelector` selector on annotations with
document URLs of the form `https://www.youtube.com/watch?v={id}`.

Fixes https://github.com/hypothesis/bouncer/issues/690